Minor bugs fixes still remaining

+++ backend/src/main/java/com/example/gov_scheme_backend/dto/response/disbursement/DisbursementMilestoneResponse.java
@@ -6,6 +6,7 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 @Data
@@ -16,11 +17,11 @@ public class DisbursementMilestoneResponse {
     private Long milestoneId;
     private Integer stageNumber;
     private String milestoneName;
-    private Double amountToRelease;
+    private BigDecimal amountToRelease;
     private LocalDate dueDate;
     private MilestoneStatus completionStatus;
     private LocalDate completedDate;
-    private Double amountReleased;
+    private BigDecimal amountReleased;
     private LocalDate releaseDate;
     private String resolvedReason;
     private LocalDate resolvedDate;
